Rinku (rinku v0.0.1)

A pattern for composing functions to execute in a chain. Execution will stop when all links in the chain have been resolved, or any link in the chain returns an error.

The initial input will be provided as the first argument in the chain. The result of each link in the chain will be supplied to the next link in the chain.

The input will always be the first argument provided to the next link in the chain.

Types

link_callback()

Specs

link_callback() ::
  (... -> any() | {:error, any()})
  | {(... -> any() | {:error, any()}), term() | list()}
  | {module(), atom(), term() | list()}

t()

Specs

t() :: %Rinku{
  links: [Rinku.Link.t()],
  resolved: [Rinku.Resolved.t()],
  result: any()
}

Functions

link(rinku, new_link, link_name \\ nil)

Specs

link(
  rinku :: t(),
  new_link :: link_callback(),
  link_name :: String.t() | atom()
) :: t()

Add a new link to the Rinku chain.

link_result(rinku, resolved_name)

Specs

link_result(rinku :: t(), link_name :: String.t()) :: any() | {:error, any()}

Get the result for a specific named execution step.

iex> Rinku.new() |> Rinku.link(fn _ -> 1 end, :step1) |> Rinku.link(fn _ -> 2 end, :step2) |> Rinku.run() |> Rinku.link_result(:step1)
1

new(intial_value \\ nil, input_name \\ :seed)

Specs

new(initial_value :: any(), input_name :: String.t() | atom()) :: t()

Create a new Rinku chain.

iex> Rinku.new()
%Rinku{
  links: [],
  resolved: [
    %Rinku.Resolved{
      name: :seed,
      result: nil
    }
  ],
  result: nil
}

result(rinku)

Specs

result(rinku :: t()) :: any() | {:error, any()}

Get the result from a processed chain.

iex> Rinku.new() |> Rinku.link(fn _ -> 1 end) |> Rinku.run() |> Rinku.result()
1

run(rinku)

Specs

run(rinku :: t()) :: t()

Execute a built rinku chain.